Refactor actions/events (#169)
* action api change

* simplify actions/events

* move logs

* event rw mutex
