+++ action.go
@@ -4,34 +4,6 @@ import (
 	"sync"
 )
 
-var (
-	// DefaultActionRegistry is the default action registry.
-	DefaultActionRegistry ActionRegistry
-)
-
-// Handle handles the named action with the given handler.
-//
-// It is a helper function that call DefaultActionRegistry.Handle.
-func Handle(name string, h ActionHandler) {
-	DefaultActionRegistry.Handle(name, h)
-}
-
-// NewAction creates and posts the named action with the given arg.
-// The action is then handled in a separate goroutine.
-//
-// It is a helper function that call DefaultActionRegistry.Post.
-func NewAction(name string, arg interface{}) {
-	DefaultActionRegistry.Post(name, arg)
-}
-
-// NewActions creates and posts a batch of actions.
-// All the actions are handled sequentially in a separate goroutine.
-//
-// It is a helper function that call DefaultActionRegistry.PostBatch.
-func NewActions(a ...Action) {
-	DefaultActionRegistry.PostBatch(a...)
-}
-
 // Action represents an action to handle.
 type Action struct {
 	Name string
@@ -41,23 +13,7 @@ type Action struct {
 // ActionHandler represent an action handler.
 type ActionHandler func(e EventDispatcher, a Action)
 
-// ActionRegistry is the interface that describes an action registry.
-type ActionRegistry interface {
-	// Handler handles the named action with the given handler.
-	Handle(name string, h ActionHandler)
-
-	// Post posts the named action with the given arg.
-	// The action is then handled in a separate goroutine.
-	Post(name string, arg interface{})
-
-	// PostBatch posts a batch of actions.
-	// All the actions are handled sequentially in a separate goroutine.
-	PostBatch(a ...Action)
-}
-
-// NewActionRegistry creates an action registry.
-// Returned registry is safe for concurrent operations.
-func NewActionRegistry(dispatcher EventDispatcher) ActionRegistry {
+func newActionRegistry(dispatcher EventDispatcher) *actionRegistry {
 	return &actionRegistry{
 		actions:    make(map[string]ActionHandler),
 		dispatcher: dispatcher,
