+++ driver.go
@@ -64,13 +64,3 @@ type Driver interface {
 
 // Addon represents a driver addon.
 type Addon func(Driver) Driver
-
-// Logs returns an addons that logs all the driver operations.
-// It uses the loggers defined in app.Loggers.
-func Logs() func(Driver) Driver {
-	return func(d Driver) Driver {
-		return &driverWithLogs{
-			Driver: d,
-		}
-	}
-}
